+++ src/components/ContactCard.astro
@@ -1,5 +1,4 @@
 ---
-import CopyIcon from "../icons/CopyIcon.astro";
 import XIcon from "../icons/XIcon.astro";
 import DownloadIcon from "../icons/DownloadIcon.astro";
 import { Image } from "astro:assets";
@@ -24,8 +23,8 @@ const contacts = [
   },
   {
     name: "Phone",
-    url: "tel:+5715985634",
-    clipboard: "+5715985634"
+    url: "https://wa.link/ou9afg",
+    clipboard: "+52 33 1793 3872"
   }
 ];
 ---
@@ -50,13 +49,15 @@ const contacts = [
         <CopyToClipboard contact={contact} client:load/>
       ))}
     </div>
-    <button class="flex items-center self-center border border-gray-700 rounded-lg w-fit px-2 py-1 text-sm bg-white font-bold text-black mt-2 hover:cursor-pointer hover:text-teal-600 transition-all duration-100 mb-4 pulse-glow hover:scale-105 hover:shadow-lg shadow-sm dark:shadow-black group">
+    <button 
+    onclick="window.open('../../src/assets/Resume.pdf', '_blank')"
+    class="flex items-center self-center border border-gray-700 rounded-lg w-fit px-2 py-1 text-sm bg-white font-bold text-black mt-2 hover:cursor-pointer hover:text-teal-600 transition-all duration-100 mb-4 pulse-glow hover:scale-105 hover:shadow-lg shadow-sm dark:shadow-black group">
       <span>Resume</span><DownloadIcon class="h-4 w-4 ml-1 dark:group-hover:text-teal-600 dark:text-black"/>
     </button>
   </div>
 </div>
 
-<script>
+<!-- <script>
   const popOverToggle = document.getElementById('pop-over-toggle');
   const profilePicture = document.getElementById('profile-picture');
   const closeButton = document.getElementById('close-button');
@@ -72,5 +73,64 @@ const contacts = [
     profilePicture?.classList.toggle('pulse');
     profilePicture?.classList.toggle('hidden');
     closeButton?.classList.toggle('hidden');
+  });
+</script> -->
+<script>
+  const popOverToggle = document.getElementById('pop-over-toggle');
+  const profilePicture = document.getElementById('profile-picture');
+  const closeButton = document.getElementById('close-button');
+  const popOver = document.getElementById('pop-over');
+
+  function openPopOver() {
+    popOver?.classList.remove('hidden');
+    popOver?.classList.add('flex');
+    popOver?.classList.remove('md:animate-pop-over-out-desktop', 'animate-pop-over-out-mobile');
+    popOver?.classList.add('md:animate-pop-over-in-desktop', 'animate-pop-over-in-mobile');
+    profilePicture?.classList.remove('pulse');
+    profilePicture?.classList.add('hidden');
+    closeButton?.classList.remove('hidden');
+  }
+
+  function closePopOver() {
+    if (popOver?.classList.contains('hidden')) return; // Already closing or hidden
+
+    popOver?.classList.remove('md:animate-pop-over-in-desktop', 'animate-pop-over-in-mobile');
+    popOver?.classList.add('md:animate-pop-over-out-desktop', 'animate-pop-over-out-mobile');
+
+    // Define a function to handle hiding after animation
+    const handleAnimationEnd = () => {
+      popOver?.classList.remove('flex');
+      popOver?.classList.add('hidden');
+      // Clean up the event listener to prevent it from firing multiple times
+      popOver?.removeEventListener('animationend', handleAnimationEnd);
+    };
+
+    // Listen for the end of the animation
+    popOver?.addEventListener('animationend', handleAnimationEnd);
+
+    profilePicture?.classList.add('pulse');
+    profilePicture?.classList.remove('hidden');
+    closeButton?.classList.add('hidden');
+  }
+
+  popOverToggle?.addEventListener('click', (event) => {
+    event.stopPropagation(); // Prevents the click from immediately closing the pop-over
+    if (popOver?.classList.contains('hidden')) {
+      openPopOver();
+    } else {
+      closePopOver();
+    }
+  });
+
+  // Listen for clicks on the whole document
+  document.addEventListener('click', (event) => {
+    // Check if the popOver is visible and the click was outside the popOver and not on the toggle button
+    if (
+      !popOver?.classList.contains('hidden') &&
+      !popOver?.contains(event.target as Node) &&
+      !popOverToggle?.contains(event.target as Node)
+    ) {
+      closePopOver();
+    }
   });
 </script>
