+++ src/components/Network.astro
@@ -46,7 +46,8 @@ import ResumeIcon from "../icons/ResumeIcon.astro";
   <!-- Resume -->
   <a
     class="relative group flex items-center hover:text-teal-500 dark:hover:text-gray-200 mr-0 hover:mr-12 transition-all duration-300 ease-out"
-    href=""
+    href="../../src/assets/Resume.pdf"
+    target="_blank"
   >
     <ResumeIcon class="h-8 w-8" />
     <span
