+++ src/layouts/Layout.astro
@@ -15,6 +15,6 @@ import { Toaster } from "@/components/ui/sonner";
 	</head>
 	<body class="font-primary" style="line-height: 1.7em;">
 		<slot />
-		<Toaster richColors={true} closeButton={true} theme='light' client:load/>
+		<Toaster richColors={true} theme='light' client:load/>
 	</body>
 </html>
